Create Script_Transfer.gs
This is an app that allows you to auto fill placeholders in a script inside of google docs with the data within a table in google sheets. 

Setup is the same as the Link_Clicker.gs, to use highlight a row and click update sheet with row and your script will update with the new information.